Add a partner file
